backend/requirements.txt
fastapi
uvicorn
prisma
openai
line-bot-sdk
stripe
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytest
httpx
pydantic[email]
jinja2
python-multipart
shiboken6
